feat: added role module mock responses in entitlement service

+++ retail/middleware/entitlement.business.service/src/app/v1/roles/roles.controller.ts
@@ -6,6 +6,10 @@ import {
   Param,
   Put,
   Delete,
+  NotFoundException,
+  ParseUUIDPipe,
+  HttpStatus,
+  HttpCode
 } from '@nestjs/common';
 import {
   ApiTags,
@@ -16,6 +20,7 @@ import {
   ApiBearerAuth,
   ApiNoContentResponse,
   ApiBadRequestResponse,
+  ApiNotFoundResponse
 } from '@nestjs/swagger';
 import { RoleService } from './roles.service';
 import { Role } from './role.entity';
@@ -28,7 +33,7 @@ export class RolesController {
   constructor(private readonly roleService: RoleService) {}
 
   @Post('/')
-  @ApiBody({ description: 'Sets the role properties.' })
+  @ApiBody({ description: 'Sets the role properties.', type: RoleDto })
   @ApiOperation({
     summary: 'Create a new role',
     description:
@@ -42,7 +47,7 @@ export class RolesController {
     type: Error,
     description: 'Input Validation failed.',
   })
-  async create(@Body() roleDto: Role): Promise<Role> {
+  async create(@Body() roleDto: RoleDto): Promise<Role> {
     return this.roleService.create(roleDto);
   }
 
@@ -67,12 +72,20 @@ export class RolesController {
     type: Role,
     description: 'Role has been successfully retrieved.',
   })
-  async findOne(@Param('id') id: string): Promise<Role> {
-    return this.roleService.findOne(id);
+  @ApiNotFoundResponse({
+    type: Error,
+    description: 'Role Not Found.',
+  })
+  async findOne(@Param('id', ParseUUIDPipe) id: string): Promise<Role> {
+    const role = await this.roleService.findOne(id);
+    if(!role) {
+      throw new NotFoundException('Role Not Found');
+    }
+    return role;
   }
 
   @Put(':id')
-  @ApiBody({ description: 'Sets the role properties.' })
+  @ApiBody({ description: 'Sets the role properties.', type: RoleDto })
   @ApiOperation({
     summary: 'Update a role by ID',
     description:
@@ -86,8 +99,16 @@ export class RolesController {
     type: Error,
     description: 'Input Validation failed.',
   })
-  async update(@Body() roleDto: Role): Promise<Role> {
-    return this.roleService.create(roleDto);
+  @ApiNotFoundResponse({
+    type: Error,
+    description: 'Role Not Found.',
+  })
+  async update(@Param('id', ParseUUIDPipe) id: string, @Body() roleDto: RoleDto): Promise<Role> {
+    const role = await this.roleService.findOne(id);
+    if(!role) {
+      throw new NotFoundException('Role Not Found');
+    }
+    return this.roleService.update(id, roleDto);
   }
 
   @Delete(':id')
@@ -97,7 +118,16 @@ export class RolesController {
       'A successful request returns the HTTP 204 No Content status code with empty response body.',
   })
   @ApiNoContentResponse({ description: 'Role has been successfully deleted.' })
-  async delete(@Param('id') id: string): Promise<any> {
-    this.roleService.delete(id);
+  @ApiNotFoundResponse({
+    type: Error,
+    description: 'Role Not Found.',
+  })
+  @HttpCode(HttpStatus.NO_CONTENT)
+  async delete(@Param('id', ParseUUIDPipe) id: string): Promise<any> {
+    const role = await this.roleService.findOne(id);
+    if(!role) {
+      throw new NotFoundException('Role Not Found');
+    }
+    return this.roleService.delete(id);
   }
 }
